refactor: extract facility list into a separate component from page
