+++ src/app/facility/search/page.tsx
@@ -1,22 +1,9 @@
 'use server';
-import FacilityListItem from './_components/FacilityListItem';
 import SearchBar from '@/app/facility/search/_components/SearchBar';
 import SearchFilterButton from '@/app/facility/search/_components/SearchFilterButton';
-import { Facility } from '@/database/schema/facility';
-
-async function getFacilities() {
-  try {
-    const res = await Facility.scan();
-    return res.Items ?? [];
-  } catch (err) {
-    console.error(err);
-    return undefined;
-  }
-}
+import FacilityList from '@/app/facility/search/_components/FacilityList';
 
 export default async function SearchPage() {
-  const facilities = await getFacilities();
-
   return (
     <>
       <SearchBar />
@@ -26,13 +13,7 @@ export default async function SearchPage() {
         <SearchFilterButton type="breastFeeding" />
         <SearchFilterButton type="familyRestroom" />
       </div>
-      <div className="flex w-full flex-col">
-        {facilities ? (
-          facilities.map((facility) => <FacilityListItem key={facility.facilityId} id={facility.facilityId} />)
-        ) : (
-          <div className="text-red-500">Error occurred while fetching data.</div>
-        )}
-      </div>
+      <FacilityList />
     </>
   );
 }
